Add automatic retry to devstack install script
I don't know about anyone else, but whenever I kick off the
ds-build role on a fresh VM, I see a high rate of failure with
git clone operations ("git call failed" is the error), which kills
off stack.sh and results in a failed build.  This happens about 50%
of the time for me, so I modified the devstack build script to
automatically retry if it encounters this error.

Change-Id: If91577707ed76b1ec566af9189e30bbdcd81c37d