Updates to devstack build profile
This patch adds the ability to specify which devstack branch to use,
adds Heat and Neutron by default (the latter replacing the deprecated
Nova networking), and corrects an intermittent glitch in the detection
of a completed stack.sh.

The documentation has also been updated, including new steps for
deploying the new version.

Change-Id: I9dbe99973be16fc2785e02ff92e052b5f37ccde9